:recycle: Separate helper to sign and validate files
